Testing a "generic" list mode for general selection
This adds a new mode `GenericListOverlay` and a new field to
`ChatState` of type `GenericListState`, which contains a bundle of
information necessary to display a simple list on the screen. This
information includes

1. the name of the list, to be displayed in the UI
2. how to render a list element (which must be a single-line
   textual string)
3. what to do when a list element is 'selected', represented as a
   callback which is provided the element in question that can
   then perform an `MH ()` computation. If the list is empty,
   then this function will never be called.
4. the actual Brick list widget

Unlike the other `ListOverlay` modes, this type is existentially
quantified over the type contained in the Brick list widget. This
means that the values contained can be anything, so long as there are
meaningful values for the associated helper functions.

Right now, there are two uses of this mode: one of them replaces the
existing behavior of the `/theme` command, which would print in the
current channel the list of possible themes; this now uses this
generic list overlay in order to select which them to use. The other
is for the `/sh` command, which used to (when used as a bare command)
simply print the list of possible `/sh`-compatible scripts; it now
lists them, and selecting one will fill in the name in your editing
window with the cursor positioned at the end, ready for the user to
continue.