async: make async actions return an optional MH action
Previously, all async actions had type 'IO (MH ())'. The IO action was
run by the async worker, and then the resulting MH () was injected into
the event loop where it was used to update the state and trigger an
application redraw. In many async cases, the IO action only ran for side
effects and then returned 'return ()'. This meant that those no-op MH
return values were still causing redraws but there was no need to do so.
This change makes it so that we can opt out of that behavior when we
know that we have no MH work to do: all async actions are now of type
'IO (Maybe (MH ()))' where a 'Nothing' means do no MH work, thus saving
a redraw.