Filter visible users based on liveness heuristic
This change modifies the user list in the sidebar so that only
non-deleted users with active DM channels appear. The arbiter for
whether they appear is the 'userShouldBeVisible' function in Types.
Spots where we only want to consider such users should use the new
'sortedVisibleUserList' function instead of 'sortedUserList', which may
eventually be phased out.