Types.Channels: stop exposing cdMessages lens, add other getters
This change adds channelMessages and modifyMessages as getter/setter
alternatives to the cdMessages lens. It also adds other getters such as
withMessages for common uses. Lastly, it adds a "dirty" flag to channels
so that any modification of the message list for a channel marks it as
dirty, and channelIsDirty and clearDirtyFlag manage the dirty state.
Forthcoming commits will use the dirty state to do cache invalidation of
rendered message lists.