Types: move all volatile ChatResources fields into a new type, MutableResources
This change creates a new record type, MutableResources, to hold all of
the data (previously) in ChatResources that is considered volatile or
mutable. The idea here (with this branch) is to get the ChatState type
to a place where it can be serialized to disk safely for the purposes
of interface draw performance analysis in a benchmarking setting. To do
that, we need to be able to selectively ignore some parts of the chat
state when we serialize it (since we can't serialize, e.g., TChans) and
when we deserialize it. In general, anything in MutableResources is
either something that can only be used in an impure (e.g. IO) setting or
something that is never, ever inspected in any drawing routines.