Add a state dump key event, config option, and command-line option
This change adds a new key event, DumpStateEvent, with a default binding
of C-\. Right now the binding is not set up using the keybinding
infrastructure because more work needs to be done to update the
keybinding conflict checker, since this binding will be top-level rather
than mode-specific, which is not a case that checker was designed to
handle.

This change also adds a config option, stateDumpLocation. It also adds a
command-line option which overrides the config file option if present.