updateViewed: take args instead of implicitly updating the current channel
This change makes updateViewed more flexible so it can be used to update
the view timestamp for any channel. It also makes the API cleaner by
taking the optional chanref for the previous channel, rather than a
boolean, to indicate whether to include the previous channel in the
request to update the view timestamp.

This change also enhances and simplifies the PostProcessMessageAdd type
and API so that it can more easily accumulate a set of added posts (to
notify the user) or chanrefs (whose view times will be updated) and
gets rid of the tricky case-by-case accumulator function that we had
before.