setFocusWith: before changing away from a channel, clear its unread status
This change is motivated by the need to clear the New Messages line from
conversation channels when switching away from them, since that is the
only mechanism we have for determining whether the channels have unread
activity. We can't use the server API call to update the channel view
timestamp the way we can with normal server channels, and that mechanism
ordinarily takes care of this concern for server channels. This change
is more general.